refactor(core): split workflow helper modules
