+++ crates/biors-core/src/workflow.rs
@@ -2,10 +2,7 @@ use crate::model_input::{
     build_model_inputs_checked, validate_model_input_policy, ModelInput, ModelInputBuildError,
     ModelInputPolicy,
 };
-use crate::sequence::{
-    validate_sequence_record, ProteinSequence, ResidueIssue, SequenceKind, SequenceRecord,
-    SequenceValidationIssue, SequenceValidationReport, ValidatedSequence,
-};
+use crate::sequence::{ProteinSequence, SequenceValidationReport};
 use crate::tokenizer::{
     protein_tokenizer_config_for_profile, summarize_tokenized_proteins,
     tokenize_protein_with_config, ProteinBatchSummary, ProteinTokenizerConfig,
@@ -13,9 +10,15 @@ use crate::tokenizer::{
 };
 use serde::{Deserialize, Serialize};
 
+mod provenance;
+mod readiness;
 mod reproducibility;
+mod validation;
+use provenance::{workflow_name, workflow_provenance};
+use readiness::readiness_issues;
 use reproducibility::{workflow_hashes, WorkflowHashInput, CORE_WORKFLOW_COMMAND};
 pub use reproducibility::{SequenceWorkflowHashes, SequenceWorkflowInvocation};
+use validation::{validate_records_for_profile, validate_workflow_input_hash};
 
 const PROTEIN_WORKFLOW_NAME: &str = "protein_model_input.v0";
 const SEQUENCE_WORKFLOW_NAME: &str = "sequence_model_input.v0";
@@ -149,127 +152,10 @@ pub fn prepare_model_input_workflow_with_config(
     Ok(SequenceWorkflowOutput {
         workflow: workflow.to_string(),
         model_ready,
-        provenance: provenance(input_hash, policy, invocation, hashes, config.profile),
+        provenance: workflow_provenance(input_hash, policy, invocation, hashes, config.profile),
         validation,
         tokenization,
         model_input,
         readiness_issues,
     })
-}
-
-fn validate_records_for_profile(
-    records: &[ProteinSequence],
-    profile: ProteinTokenizerProfile,
-) -> SequenceValidationReport {
-    let kind = profile.sequence_kind();
-    let sequences = records
-        .iter()
-        .map(|record| {
-            let validation = validate_sequence_record(&SequenceRecord {
-                id: record.id.clone(),
-                sequence: String::from_utf8_lossy(&record.sequence).into_owned(),
-                kind,
-            });
-            ValidatedSequence {
-                id: validation.id,
-                sequence: validation.sequence,
-                alphabet: validation.alphabet,
-                valid: validation.valid,
-                warnings: validation
-                    .warnings
-                    .into_iter()
-                    .map(issue_to_residue_issue)
-                    .collect(),
-                errors: validation
-                    .errors
-                    .into_iter()
-                    .map(issue_to_residue_issue)
-                    .collect(),
-            }
-        })
-        .collect();
-    crate::sequence::summarize_validated_sequences(sequences)
-}
-
-fn issue_to_residue_issue(issue: SequenceValidationIssue) -> ResidueIssue {
-    ResidueIssue {
-        residue: issue.symbol,
-        position: issue.position,
-    }
-}
-
-fn validate_workflow_input_hash(input_hash: &str) -> Result<(), ModelInputBuildError> {
-    if crate::verification::is_stable_input_hash(input_hash) {
-        Ok(())
-    } else {
-        Err(ModelInputBuildError::InvalidInputHash {
-            input_hash: input_hash.to_string(),
-        })
-    }
-}
-
-fn provenance(
-    input_hash: String,
-    policy: ModelInputPolicy,
-    invocation: SequenceWorkflowInvocation,
-    hashes: SequenceWorkflowHashes,
-    profile: ProteinTokenizerProfile,
-) -> SequenceWorkflowProvenance {
-    let vocab = crate::tokenizer::inspect_protein_tokenizer_config(
-        &protein_tokenizer_config_for_profile(profile),
-    )
-    .vocabulary;
-    SequenceWorkflowProvenance {
-        biors_core_version: env!("CARGO_PKG_VERSION").to_string(),
-        invocation,
-        input_hash,
-        normalization: NORMALIZATION_POLICY.to_string(),
-        validation_alphabet: vocab.name.clone(),
-        tokenizer: WorkflowTokenizerMetadata {
-            name: vocab.name.clone(),
-            vocab_size: vocab.tokens.len(),
-            unknown_token_id: vocab.unknown_token_id,
-            unknown_token_policy: vocab.unknown_token_policy.clone(),
-        },
-        model_input_policy: policy,
-        hashes,
-    }
-}
-
-fn workflow_name(profile: ProteinTokenizerProfile) -> &'static str {
-    match profile.sequence_kind() {
-        SequenceKind::Protein => PROTEIN_WORKFLOW_NAME,
-        SequenceKind::Dna | SequenceKind::Rna => SEQUENCE_WORKFLOW_NAME,
-    }
-}
-
-fn readiness_issues(tokenized: &[TokenizedProtein]) -> Vec<SequenceWorkflowReadinessIssue> {
-    tokenized
-        .iter()
-        .filter(|record| {
-            record.tokens.is_empty() || !record.warnings.is_empty() || !record.errors.is_empty()
-        })
-        .map(|record| {
-            let warning_count = record.warnings.len();
-            let error_count = record.errors.len();
-            let message = if record.tokens.is_empty() {
-                format!(
-                    "sequence '{}' is not model-ready: empty sequences cannot be converted into model input",
-                    record.id
-                )
-            } else {
-                format!(
-                    "sequence '{}' is not model-ready: {warning_count} warnings and {error_count} errors must be resolved before model-input generation",
-                    record.id
-                )
-            };
-            SequenceWorkflowReadinessIssue {
-                code: READINESS_ISSUE_CODE.to_string(),
-                id: record.id.clone(),
-                warning_count,
-                error_count,
-                message,
-            }
-        })
-        .collect()
 }
